LITE-22101 Prepared separation of Django-RQL to 2 libraries
