+++ dj_rql/drf/paginations.py
@@ -1,13 +1,14 @@
 #
-#  Copyright © 2021 Ingram Micro Inc. All rights reserved.
+#  Copyright © 2022 Ingram Micro Inc. All rights reserved.
 #
 
 from dj_rql.drf._utils import get_query
-from dj_rql.exceptions import RQLFilterParsingError
-from dj_rql.parser import RQLParser
 from dj_rql.transformer import RQLLimitOffsetTransformer
 
 from lark.exceptions import LarkError
+
+from py_rql.exceptions import RQLFilterParsingError
+from py_rql.parser import RQLParser
 
 from rest_framework.pagination import LimitOffsetPagination, _positive_int
 from rest_framework.response import Response
