--- tests/test_parser/__init__.py
@@ -1,3 +1,0 @@
-#
-#  Copyright © 2020 Ingram Micro Inc. All rights reserved.
-#
