+++ tests/test_py_rql/test_parser/constants.py
@@ -1,8 +1,8 @@
 #
-#  Copyright © 2021 Ingram Micro Inc. All rights reserved.
+#  Copyright © 2022 Ingram Micro Inc. All rights reserved.
 #
 
-from dj_rql.constants import (
+from py_rql.constants import (
     ComparisonOperators, ListOperators, LogicalOperators, RQL_EMPTY, RQL_NULL, SearchOperators,
 )
