+++ tests/test_py_rql/test_parser/utils.py
@@ -2,7 +2,7 @@
 #  Copyright © 2020 Ingram Micro Inc. All rights reserved.
 #
 
-from dj_rql.transformer import BaseRQLTransformer
+from py_rql.transformer import BaseRQLTransformer
 
 
 class ComparisonTransformer(BaseRQLTransformer):
